Review examples fixes
