+++ configs/router.conf
@@ -1,6 +1,36 @@
+log-search = {
+  path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
+  className = "SimpleTextSearch$"
+  namespace = "foo"
+}
+
+streaming-log-search = {
+  path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
+  className = "StreamingTextSearch$"
+  namespace = "streaming"
+}
+
+classification = {
+  path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
+  className = "MLClassification$"
+  namespace = "regression"
+}
+
+simple-context = {
+  path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
+  className = "SimpleContext$"
+  namespace = "foo"
+}
+
+simple-context-py = {
+  path = "./examples/src/main/python/simple_context.py"
+  className = "SimpleContext"
+  namespace = "foo"
+}
+
 dtree {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "DTree$"
+  className = "DTreeClassificationJob$"
   namespace = "DTree"
 }
 
@@ -12,42 +42,42 @@ regression {
 
 random_forest {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "RandomForest$"
+  className = "RandomForestClassificationJob$"
   namespace = "RandomForest"
 }
 
-binarizer {
-  path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "Binarizer$"
-  namespace = "Binarizer"
-}
+ binarizer {
+   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
+   className = "BinarizerJob$"
+   namespace = "Binarizer"
+ }
 
 pca {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "PCA$"
+  className = "PCAJob$"
   namespace = "PCA"
 }
 
 minmaxscaler {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "MinMaxScaler$"
+  className = "MinMaxScalerJob$"
   namespace = "MinMaxScaler"
 }
 
 standardscaler {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "StandardScaler$"
+  className = "StandardScalerJob$"
   namespace = "StandardScaler"
 }
 
 maxabsscaler {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "MaxAbsScaler$"
+  className = "MaxAbsScalerJob$"
   namespace = "MaxAbsScaler"
 }
 
 stringindexer {
   path = "./examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar"
-  className = "StringIndexer$"
+  className = "StringIndexerJob$"
   namespace = "StringIndexer"
 }
