+++ examples/src/main/scala/BinarizerJob_Spark2.scala
@@ -3,9 +3,8 @@ import org.apache.spark.ml.Pipeline
 import org.apache.spark.ml.feature.Binarizer
 
 
-object Binarizer extends MLMistJob with SQLSupport {
+object BinarizerJob extends MLMistJob with SQLSupport {
   def train(savePath: String): Map[String, Any] = {
-    assert(savePath == "src/test/resources/models/binarizer")
     val data = Array((0, 0.1), (1, 0.8), (2, 0.2))
     val dataFrame = session.createDataFrame(data).toDF("id", "feature")
 
@@ -18,7 +17,6 @@ object Binarizer extends MLMistJob with SQLSupport {
 
     val model = pipeline.fit(dataFrame)
 
-    //model.write.overwrite().save("src/test/resources/models/binarizer")
     model.write.overwrite().save(savePath)
     Map.empty[String, Any]
   }
