+++ examples/src/main/scala/DTreeClassificationJob_Spark2.scala
@@ -1,14 +1,10 @@
 import io.hydrosphere.mist.lib._
 import org.apache.spark.ml.Pipeline
-import org.apache.spark.ml.classification.{RandomForestClassificationModel, RandomForestClassifier}
-import org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator
+import org.apache.spark.ml.classification.DecisionTreeClassifier
 import org.apache.spark.ml.feature.{IndexToString, StringIndexer, VectorIndexer}
-import org.apache.spark.ml.linalg.{Vector, Vectors}
+import org.apache.spark.ml.linalg.{DenseVector, SparseVector, Vector, Vectors}
 
-/**
-  * Created by Bulat on 22.02.2017.
-  */
-object RandomForest extends MLMistJob with SQLSupport {
+object DTreeClassificationJob extends MLMistJob with SQLSupport {
   def constructVector(params: Map[String, Any]): Vector = {
     Vectors.sparse(
       params("size").asInstanceOf[Int],
@@ -17,50 +13,35 @@ object RandomForest extends MLMistJob with SQLSupport {
     )
   }
 
-  def train(savePath: String, datasetPath: String): Map[String, Any] = {
-    // Load and parse the data file, converting it to a DataFrame.
+  def train(datasetPath: String, savePath: String): Map[String, Any] = {
     val data = session.read.format("libsvm").load(datasetPath)
-
-    // Index labels, adding metadata to the label column.
-    // Fit on whole dataset to include all labels in index.
+    val Array(training, _) = data.randomSplit(Array(0.7, 0.3))
     val labelIndexer = new StringIndexer()
       .setInputCol("label")
       .setOutputCol("indexedLabel")
       .fit(data)
-    // Automatically identify categorical features, and index them.
-    // Set maxCategories so features with > 4 distinct values are treated as continuous.
     val featureIndexer = new VectorIndexer()
       .setInputCol("features")
       .setOutputCol("indexedFeatures")
-      .setMaxCategories(4)
+      .setMaxCategories(4)// features with > 4 distinct values are treated as continuous.
       .fit(data)
-
-    // Split the data into training and test sets (30% held out for testing).
-    val Array(trainingData, testData) = data.randomSplit(Array(0.7, 0.3))
-
-    // Train a RandomForest model.
-    val rf = new RandomForestClassifier()
+    val dt = new DecisionTreeClassifier()
       .setLabelCol("indexedLabel")
       .setFeaturesCol("indexedFeatures")
-      .setNumTrees(10)
 
-    // Convert indexed labels back to original labels.
     val labelConverter = new IndexToString()
       .setInputCol("prediction")
       .setOutputCol("predictedLabel")
       .setLabels(labelIndexer.labels)
 
-    // Chain indexers and forest in a Pipeline.
     val pipeline = new Pipeline()
-      .setStages(Array(labelIndexer, featureIndexer, rf, labelConverter))
+      .setStages(Array(labelIndexer, featureIndexer, dt, labelConverter))
 
-    // Train model. This also runs the indexers.
-    val model = pipeline.fit(trainingData)
+    val model = pipeline.fit(training)
 
     model.write.overwrite().save(savePath)
     Map.empty[String, Any]
-  }
-
+}
   def serve(modelPath: String, features: Map[String, Any]): Map[String, Any] = {
     import io.hydrosphere.mist.ml.transformers.LocalTransformers._
 
@@ -68,7 +49,6 @@ object RandomForest extends MLMistJob with SQLSupport {
     val data = LocalData(
       LocalDataColumn("features", List(constructVector(features)))
     )
-
     val result: LocalData = pipeline.transform(data)
     Map("result" -> result.select("predictedLabel").toMapList)
   }
