+++ examples/src/main/scala/MaxAbsScalerJob_Spark2.scala
@@ -1,29 +1,25 @@
 import io.hydrosphere.mist.lib._
+
 import org.apache.spark.ml.Pipeline
-import org.apache.spark.ml.feature.StandardScaler
+import org.apache.spark.ml.feature.MaxAbsScaler
 import org.apache.spark.ml.linalg.Vectors
 
 
-object StandardScaler extends MLMistJob with SQLSupport {
+object MaxAbsScalerJob extends MLMistJob with SQLSupport {
   def train(savePath: String): Map[String, Any] = {
-    assert(savePath == "src/test/resources/models/standardscaler")
+    val dataFrame = session.createDataFrame(Seq(
+      (0, Vectors.dense(1.0, 0.1, -8.0)),
+      (1, Vectors.dense(2.0, 1.0, -4.0)),
+      (2, Vectors.dense(4.0, 10.0, 8.0))
+    )).toDF("id", "features")
 
-    val data = Array(
-      Vectors.dense(0.0, 10.3, 1.0, 4.0, 5.0),
-      Vectors.dense(2.0, 0.0, 3.0, 4.0, 5.0),
-      Vectors.dense(4.0, 0.0, 0.0, 6.0, 7.0)
-    )
-    val df = session.createDataFrame(data.map(Tuple1.apply)).toDF("features")
-
-    val scaler = new StandardScaler()
+    val scaler = new MaxAbsScaler()
       .setInputCol("features")
       .setOutputCol("scaledFeatures")
-      .setWithStd(true)
-      .setWithMean(false)
 
     val pipeline = new Pipeline().setStages(Array(scaler))
 
-    val model = pipeline.fit(df)
+    val model = pipeline.fit(dataFrame)
 
     model.write.overwrite().save(savePath)
     Map.empty[String, Any]
@@ -38,6 +34,6 @@ object StandardScaler extends MLMistJob with SQLSupport {
     )
 
     val result: LocalData = pipeline.transform(data)
-    Map("result" -> result.select("features", "scaledFeatures").toMapList)
+    Map("result" -> result.select("scaledFeatures").toMapList)
   }
 }
