+++ examples/src/main/scala/RandomForestClassificationJob_Spark2.scala
@@ -1,49 +1,63 @@
 import io.hydrosphere.mist.lib._
 import org.apache.spark.ml.Pipeline
-import org.apache.spark.ml.classification.DecisionTreeClassifier
+import org.apache.spark.ml.classification.{RandomForestClassificationModel, RandomForestClassifier}
 import org.apache.spark.ml.feature.{IndexToString, StringIndexer, VectorIndexer}
-import org.apache.spark.ml.linalg.{DenseVector, SparseVector, Vector, Vectors}
+import org.apache.spark.ml.linalg.{Vector, Vectors}
 
-object DTree extends MLMistJob with SQLSupport {
+
+object RandomForestClassificationJob extends MLMistJob with SQLSupport {
   def constructVector(params: Map[String, Any]): Vector = {
     Vectors.sparse(
       params("size").asInstanceOf[Int],
       params("indices").asInstanceOf[List[Int]].toArray[Int],
-      params("values").asInstanceOf[List[Int]].map(_.toDouble).toArray[Double] // why Int? I pass Double though
+      params("values").asInstanceOf[List[Int]].map(_.toDouble).toArray[Double]
     )
   }
 
-  def train(datasetPath: String, savePath: String): Map[String, Any] = {
-    //val data = session.read.format("libsvm").load("/home/bulat/Documents/spark-2.1.0-bin-hadoop2.7//data/mllib/sample_libsvm_data.txt")
+  def train(savePath: String, datasetPath: String): Map[String, Any] = {
+    // Load and parse the data file, converting it to a DataFrame.
     val data = session.read.format("libsvm").load(datasetPath)
-    val Array(training, _) = data.randomSplit(Array(0.7, 0.3))
+
+    // Index labels, adding metadata to the label column.
+    // Fit on whole dataset to include all labels in index.
     val labelIndexer = new StringIndexer()
       .setInputCol("label")
       .setOutputCol("indexedLabel")
       .fit(data)
+    // Automatically identify categorical features, and index them.
+    // Set maxCategories so features with > 4 distinct values are treated as continuous.
     val featureIndexer = new VectorIndexer()
       .setInputCol("features")
       .setOutputCol("indexedFeatures")
-      .setMaxCategories(4)// features with > 4 distinct values are treated as continuous.
+      .setMaxCategories(4)
       .fit(data)
-    val dt = new DecisionTreeClassifier()
+
+    // Split the data into training and test sets (30% held out for testing).
+    val Array(trainingData, testData) = data.randomSplit(Array(0.7, 0.3))
+
+    // Train a RandomForest model.
+    val rf = new RandomForestClassifier()
       .setLabelCol("indexedLabel")
       .setFeaturesCol("indexedFeatures")
+      .setNumTrees(10)
 
+    // Convert indexed labels back to original labels.
     val labelConverter = new IndexToString()
       .setInputCol("prediction")
       .setOutputCol("predictedLabel")
       .setLabels(labelIndexer.labels)
 
+    // Chain indexers and forest in a Pipeline.
     val pipeline = new Pipeline()
-      .setStages(Array(labelIndexer, featureIndexer, dt, labelConverter))
+      .setStages(Array(labelIndexer, featureIndexer, rf, labelConverter))
 
-    val model = pipeline.fit(training)
+    // Train model. This also runs the indexers.
+    val model = pipeline.fit(trainingData)
 
-    //model.write.overwrite().save("models/dtree")
     model.write.overwrite().save(savePath)
     Map.empty[String, Any]
-}
+  }
+
   def serve(modelPath: String, features: Map[String, Any]): Map[String, Any] = {
     import io.hydrosphere.mist.ml.transformers.LocalTransformers._
 
@@ -51,6 +65,7 @@ object DTree extends MLMistJob with SQLSupport {
     val data = LocalData(
       LocalDataColumn("features", List(constructVector(features)))
     )
+
     val result: LocalData = pipeline.transform(data)
     Map("result" -> result.select("predictedLabel").toMapList)
   }
