+++ examples/src/main/scala/StandardScalerJob_Spark2.scala
@@ -1,27 +1,27 @@
 import io.hydrosphere.mist.lib._
-
 import org.apache.spark.ml.Pipeline
-import org.apache.spark.ml.feature.MinMaxScaler
+import org.apache.spark.ml.feature.StandardScaler
 import org.apache.spark.ml.linalg.Vectors
 
 
-object MinMaxScaler extends MLMistJob with SQLSupport {
+object StandardScalerJob extends MLMistJob with SQLSupport {
   def train(savePath: String): Map[String, Any] = {
-    assert(savePath == "src/test/resources/models/minmaxscaler")
+    val data = Array(
+      Vectors.dense(0.0, 10.3, 1.0, 4.0, 5.0),
+      Vectors.dense(2.0, 0.0, 3.0, 4.0, 5.0),
+      Vectors.dense(4.0, 0.0, 0.0, 6.0, 7.0)
+    )
+    val df = session.createDataFrame(data.map(Tuple1.apply)).toDF("features")
 
-    val dataFrame = session.createDataFrame(Seq(
-      (0, Vectors.dense(1.0, 0.1, -1.0)),
-      (1, Vectors.dense(2.0, 1.1, 1.0)),
-      (2, Vectors.dense(3.0, 10.1, 3.0))
-    )).toDF("id", "features")
-
-    val scaler = new MinMaxScaler()
+    val scaler = new StandardScaler()
       .setInputCol("features")
       .setOutputCol("scaledFeatures")
+      .setWithStd(true)
+      .setWithMean(false)
 
     val pipeline = new Pipeline().setStages(Array(scaler))
 
-    val model = pipeline.fit(dataFrame)
+    val model = pipeline.fit(df)
 
     model.write.overwrite().save(savePath)
     Map.empty[String, Any]
@@ -36,6 +36,6 @@ object MinMaxScaler extends MLMistJob with SQLSupport {
     )
 
     val result: LocalData = pipeline.transform(data)
-    Map("result" -> result.select("scaledFeatures").toMapList)
+    Map("result" -> result.select("features", "scaledFeatures").toMapList)
   }
 }
