+++ src/test/resources/reference.conf
@@ -312,7 +312,7 @@ mist.test.request.restificated = """
 mist.test.ml.binarizer = """
   {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "Binarizer$",
+    "className": "BinarizerJob$",
     "parameters": {
       "modelPath": "src/test/resources/models/binarizer",
       "features": [0.1, 0.1, 5.9]
@@ -324,7 +324,7 @@ mist.test.ml.binarizer = """
 mist.test.ml.pca = """
   {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "PCA$",
+    "className": "PCAJob$",
     "parameters": {
       "modelPath": "src/test/resources/models/pca",
       "features": [[2.0, 0.0, 3.0, 4.0, 5.0], [4.0, 0.0, 0.0, 6.0, 7.0]]
@@ -336,7 +336,7 @@ mist.test.ml.pca = """
 mist.test.ml.standardscaler = """
   {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "StandardScaler$",
+    "className": "StandardScalerJob$",
     "parameters": {
       "modelPath": "src/test/resources/models/standardscaler",
       "features": [[1.0, 0.0, 1.0, 2.0, 0.0], [2.0, 0.0, 3.0, 4.0, 5.0], [4.0, 0.0, 0.0, 6.0, 7.0]]
@@ -348,7 +348,7 @@ mist.test.ml.standardscaler = """
 mist.test.ml.minmaxscaler = """
   {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "MinMaxScaler$",
+    "className": "MinMaxScalerJob$",
     "parameters": {
       "modelPath": "src/test/resources/models/minmaxscaler",
       "features": [[1.0, 0.0, -1.0], [2.0, 1.5, 1.0], [3.0, 1.1, 3.0]]
@@ -360,7 +360,7 @@ mist.test.ml.minmaxscaler = """
 mist.test.ml.maxabsscaler = """
   {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "MaxAbsScaler$",
+    "className": "MaxAbsScalerJob$",
     "parameters": {
       "modelPath": "src/test/resources/models/maxabsscaler",
       "features": [[1.0, 0.0, 1.0], [2.0, 4.0, 5.0], [0.0, 6.0, 7.0]]
@@ -372,7 +372,7 @@ mist.test.ml.maxabsscaler = """
 mist.test.ml.stringindexer = """
   {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "StringIndexer$",
+    "className": "StringIndexerJob$",
     "parameters": {
       "modelPath": "src/test/resources/models/stringindexer",
       "features": ["a", "b", "c"]
@@ -385,7 +385,7 @@ mist.test.ml.stringindexer = """
 mist.test.ml.tree-classifier_1 = """
   {
       "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-      "className": "DTree$",
+      "className": "DTreeClassificationJob$",
       "parameters": {
           "modelPath": "src/test/resources/models/tree-classifier",
           "features": {
@@ -439,7 +439,7 @@ mist.test.ml.tree-classifier_1 = """
 mist.test.ml.tree-classifier_0 = """
   {
       "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-      "className": "DTree$",
+      "className": "DTreeClassificationJob$",
       "parameters": {
               "modelPath": "src/test/resources/models/tree-classifier",
               "features": {
@@ -540,7 +540,7 @@ mist.test.ml.tree-classifier_0 = """
 mist.test.ml.forest-classifier_1 = """
 {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "RandomForest$",
+    "className": "RandomForestClassificationJob$",
     "parameters": {
         "modelPath": "src/test/resources/models/forest-classifier",
         "features": {
@@ -574,7 +574,7 @@ mist.test.ml.forest-classifier_1 = """
 mist.test.ml.forest-classifier_0 = """
 {
     "path": "examples/target/scala-2.11/mist_examples_2.11-0.8.0.jar",
-    "className": "RandomForest$",
+    "className": "RandomForestClassificationJob$",
     "parameters": {
         "modelPath": "src/test/resources/models/forest-classifier",
         "features": {
